FilterableDropdown: DRY (shared abstract class)
