+++ WolvenKit/Views/Templates/FilterableDropdown/FilterableDropdownCNameMenu.xaml
@@ -1,6 +1,6 @@
-﻿<reactiveUi:ReactiveUserControl
+﻿<templates:FilterableDropdownMenuBase
     x:Class="WolvenKit.Views.Editors.FilterableDropdownCNameMenu"
-    x:TypeArguments="shell:ChunkViewModel"
+    x:TypeArguments="types:CName"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:helpers="clr-namespace:WolvenKit.Functionality.Helpers"
@@ -10,6 +10,7 @@
     xmlns:shell="clr-namespace:WolvenKit.App.ViewModels.Shell;assembly=WolvenKit.App"
     xmlns:reactiveUi="http://reactiveui.net"
     xmlns:templates="clr-namespace:WolvenKit.Views.Templates"
+    xmlns:types="clr-namespace:WolvenKit.RED4.Types;assembly=WolvenKit.RED4"
     FocusManager.FocusedElement="{Binding ElementName=FilterTextBox}">
     <Grid>
         <Grid.Resources>
@@ -149,4 +150,4 @@
             </editors:RedCNameEditor.Style>
         </editors:RedCNameEditor>
     </Grid>
-</reactiveUi:ReactiveUserControl>
+</templates:FilterableDropdownMenuBase>
