+++ WolvenKit/Views/Templates/FilterableDropdown/FilterableDropdownMenu.xaml.cs
