+++ WolvenKit/Views/Templates/FilterableDropdown/FilterableDropdownRedstringMenu.xaml
@@ -1,6 +1,6 @@
-﻿<reactiveUi:ReactiveUserControl
-    x:Class="WolvenKit.Views.Editors.FilterableDropdownResourcePathMenu"
-    x:TypeArguments="shell:ChunkViewModel"
+﻿<templates:FilterableDropdownMenuBase
+    x:Class="WolvenKit.Views.Editors.FilterableDropdownRedstringMenu"
+    x:TypeArguments="types:IRedString"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:helpers="clr-namespace:WolvenKit.Functionality.Helpers"
@@ -10,6 +10,7 @@
     xmlns:shell="clr-namespace:WolvenKit.App.ViewModels.Shell;assembly=WolvenKit.App"
     xmlns:reactiveUi="http://reactiveui.net"
     xmlns:templates="clr-namespace:WolvenKit.Views.Templates"
+    xmlns:types="clr-namespace:WolvenKit.RED4.Types;assembly=WolvenKit.RED4"
     FocusManager.FocusedElement="{Binding ElementName=FilterTextBox}">
     <Grid>
         <Grid.Resources>
@@ -18,20 +19,27 @@
                     <hc:ThemeResources />
                     <hc:Theme />
                 </ResourceDictionary.MergedDictionaries>
+                <Style
+                    BasedOn="{StaticResource WolvenkitDialogLabelStyleMedium}"
+                    TargetType="{x:Type Label}" />
+                <Style TargetType="{x:Type Grid}">
+                    <Setter Property="Background" Value="{StaticResource BackgroundColor_Control}" />
+                </Style>
                 <converters:IntToVisibilityConverterInverted x:Key="IntToVisibilityConverterInverted" />
-                <converters:IRedTypeToCNameConverter x:Key="IRedTypeToCNameConverter" />
+                <converters:IRedTypeToCNameConverter x:Key="IRedTypeToRedstringConverter" />
             </ResourceDictionary>
         </Grid.Resources>
 
         <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*" MinWidth="{StaticResource WolvenKitPluginsToolIconButtonSize}" MaxWidth="{StaticResource WolvenKitLabelWidthLarge}" />
+            <ColumnDefinition Width="*" MinWidth="{StaticResource WolvenKitLabelWidth}" />
             <ColumnDefinition Width="*" MinWidth="{StaticResource WolvenKitLabelWidth}" />
             <ColumnDefinition x:Name="Col3" Width="{StaticResource WolvenKitButtonGridWidth}" />
         </Grid.ColumnDefinitions>
         <Grid.RowDefinitions>
-            <RowDefinition x:Name="FilterRow" Height="Auto" />
-            <RowDefinition x:Name="PlaceholderRow" Height="5" />
+            <RowDefinition x:Name="Row1" Height="Auto" />
+            <RowDefinition x:Name="Row2" Height="5" />
             <RowDefinition Height="Auto" />
+            <RowDefinition Height="5" />
         </Grid.RowDefinitions>
 
         <!-- ====================================== -->
@@ -47,8 +55,9 @@
             Style="{StaticResource WolvenkitTextBoxStyle}"
             helpers:TextBoxBehavior.TripleClickSelectAll="True"
             KeyboardNavigation.TabIndex="1"
+            KeyDown="FilterTextBox_KeyDown"
             Text="{Binding FilterText,
-                           RelativeSource={RelativeSource AncestorType={x:Type editors:FilterableDropdownResourcePathMenu}},
+                           RelativeSource={RelativeSource AncestorType={x:Type editors:FilterableDropdownRedstringMenu}},
                            Mode=TwoWay}" />
         <!-- Placeholder text -->
         <TextBlock
@@ -70,13 +79,13 @@
             Grid.Column="1"
             Height="{StaticResource WolvenKitTabHeight}"
             ItemsSource="{Binding FilteredOptions,
-                                  RelativeSource={RelativeSource AncestorType={x:Type editors:FilterableDropdownResourcePathMenu}},
+                                  RelativeSource={RelativeSource AncestorType={x:Type editors:FilterableDropdownRedstringMenu}},
                                   Mode=OneWay}"
             KeyboardNavigation.TabIndex="2"
             SelectedValuePath="Value"
             DisplayMemberPath="Key"
             SelectedValue="{Binding SelectedOption,
-                                    RelativeSource={RelativeSource AncestorType={x:Type editors:FilterableDropdownResourcePathMenu}},
+                                    RelativeSource={RelativeSource AncestorType={x:Type editors:FilterableDropdownRedstringMenu}},
                                     Mode=TwoWay}" />
 
         <!-- ====================================== -->
@@ -102,14 +111,23 @@
         </Button>
 
         <!-- ====================================== -->
-        <!-- CName editor -->
+        <!-- Redstring editor -->
         <!-- ====================================== -->
-        <editors:RedRefEditor
-            x:Name="RedRefEditor"
+
+        <editors:RedStringEditor
+            x:Name="RedStringEditor"
             Grid.Row="2"
             Grid.Column="0"
             Grid.ColumnSpan="2"
-            RedRef="{Binding Data,
-                             Mode=TwoWay}" />
+            RedString="{Binding Data,
+                                Mode=TwoWay}">
+            <editors:RedStringEditor.Style>
+                <Style TargetType="editors:RedStringEditor">
+                    <EventSetter
+                        Event="ValueChanged"
+                        Handler="RedStringEditor_ValueChanged" />
+                </Style>
+            </editors:RedStringEditor.Style>
+        </editors:RedStringEditor>
     </Grid>
-</reactiveUi:ReactiveUserControl>
+</templates:FilterableDropdownMenuBase>
